+++ firebase-vscode/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## NEXT
 
+- IDX Auth is picked up by VSCode
 - [Fixed] Data Connect emulator issues properly streamed on startup
 - [Fixed] Data Connect schema reloads consistently